+++ app/config/versions.js
@@ -1,30 +1,24 @@
 module.exports = {
     libs: {
-        'bem-core': 'v2.6.0',
-        'bem-components': 'v2.2.1'
+        'bem-core': 'v2.7.0',
+        'bem-components': 'v2.3.0'
     },
     deps: {
         // enb
-        enb: '0.15.0',
-        'enb-autoprefixer': '0.3.0',
-        'enb-bemxjst': '1.3.4',
-        'enb-bem-techs': '1.0.4',
-        'enb-bh': '0.5.0',
-        'enb-borschik': '1.5.0',
-        'enb-diverse-js': '0.1.0',
-        'enb-modules': '0.2.0',
-        'enb-stylus': '1.2.3',
+        enb: '^0.17.0',
+        'enb-autoprefixer': '^0.3.0',
+        'enb-bemxjst': '^1.3.5',
+        'enb-bem-techs': '^2.0.0-rc',
+        'enb-bh': '^1.0.0-rc',
+        'enb-borschik': '^1.5.1',
+        'enb-diverse-js': '^0.1.0',
+        'enb-modules': '^0.2.0',
+        'enb-stylus': '^2.0.0-rc',
 
         // bem-tools
-        bem: '0.9.0',
-        'bem-environ': '1.4.0',
-        'bem-tools-autoprefixer': '0.0.3',
+        bem: '^0.10.0',
 
-        // other
-        bh: '3.3.0',
-        bower: '1.4.1',
-        'borschik-tech-cleancss': '2.0.1',
-        'bower-npm-install': '0.5.9',
-        stylus: '0.50.0'
+        bower: '^1.4.1',
+        'borschik-tech-cleancss': '^2.0.1'
     }
 };