+++ app/templates/.bem/techs/bemjson.js.js
@@ -6,27 +6,27 @@ exports.API_VER = 2;
 
 exports.techMixin = {
 
-	getCreateResult : function(path, suffix, vars) {
+    getCreateResult : function(path, suffix, vars) {
 
-		if (vars.opts && vars.opts.content) return vars.opts.content;
+        if (vars.opts && vars.opts.content) return vars.opts.content;
 
-		return Template.process([
-			"({",
-			"	block: 'page',",
-			"	title: '{{bemBlockName}}',",
-			"	head: [",
-			"		{ elem: 'css', url: '_{{bemBlockName}}.css' }",
-			"	],",
-			"	scripts: [{ elem: 'js', url: '_{{bemBlockName}}.js' }],",
-			"	content: [",
-			"		{",
-			"			block: 'content',",
-			"			content: [",
-			"				'block content'",
-			"			]",
-			"		}",
-			"	]",
-			"})"], vars);
-	}
+        return Template.process([
+            "module.exports = {",
+            "    block: 'page',",
+            "    title: '{{bemBlockName}}',",
+            "    head: [",
+            "        { elem: 'css', url: '_{{bemBlockName}}.css' }",
+            "    ],",
+            "    scripts: [{ elem: 'js', url: '_{{bemBlockName}}.js' }],",
+            "    content: [",
+            "       {",
+            "           block: 'content',",
+            "           content: [",
+            "               'block content'",
+            "           ]",
+            "       }",
+            "    ]",
+            "};"], vars);
+    }
 
 };